packages/github-integration/pyproject.toml
[project]
name = "eudorax-github-integration"
version = "1.0.0"
description = "GitHub PR optimization and learning for EudoraX"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "eudorax-core",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
